move cache utils into cache module
