+++ src/lib.rs
@@ -4,4 +4,3 @@ pub mod metrics_server;
 pub mod mojang;
 pub mod service;
 pub mod settings;
-pub mod util;
